Added thread slug customization support
Hashover adds a thread slug with each comment. Some static site generators create urls with date. So it causes it to not show the comments at all. This commit gives it a chance to modify configs and make it so that it works on them.

Config options are added for both pages and posts, so any destination url structure should work fine now.